Service registration
Service is the interface used to expose a service a user can enable/disable
in the server/clients.

Check method verified if Message has the expected properties.
Process execute the behavior that message needs to get done.

Services.Register is a way to map a service to a route name and resolve if a message "Kind"
has a service to handle with It.

If has no service to handle a message, ti will be processed with the default service.
The default service check the message as false with the error "unsuported message".